Mahindra Panvel Mumbai
Mahindra Lifespaces has now launched its latest project titled Mahindra Panvel Mumbai, the project has been designed to provide an abode of comfortable living with all the world-class amenities. It is ideal for buyers looking for stylish and premium real estate residences at reasonable prices. The project will have inventory configurations comprising 2 & 3 BHK apartments accompanied by all the basic and leisure amenities.
Upcoming Mahindra Panvel offers a host of locational advantages such as close proximity to major IT organisations, hospitals, prominent schools and colleges and also the Orion Mall, and Seawoods Grand Central Mall. The Panvel railway station is just a few km away from the property. It is also well-linked with the Mumbai-Pune Highway, NH-48, and NH-66 which greatly increases the connectivity routes of the project. Moreover, the project is located in the heart of Mumbai and enjoys all the benefits the city has to offer. 
The development is equipped with all the necessary modern amenities and the latest multi-level security systems. The ambience is covered with ample trees and parks to provide a breath of fresh air and improve the lifestyle of its residents. Amongst the many new residential projects in Mumbai, Mahindra Panvel Flats has proved itself to be a worthy project by providing a premium real estate home with all the world-class amenities at reasonable prices.
Unwind yourself with a leisurely walk along a covered pathway surrounded by trees. Spend time with young ones in dedicated play areas, swim in an Olympic-size pool, and shop for daily necessities just around the corner. 
About Builder
Mahindra Lifespaces is one of the most prominent real estate developers in India and has been serving the real estate sector since 1994. They are known to deliver some of the most exclusive residential spaces in Mumbai city. The group is committed to eco-friendly construction and innovative architectural designs. The group has also been bestowed with an A+ rating by the Government of Maharashtra for its excellent work.
To know more - https://bit.ly/3hyNPMp